Enhanced tests and fix for #7668 - re-adding a component at an index before its current one does nothing
Simplified the logic in AbstractOrderedLayout/CssLayout#addComponent*() - super.addComponent() is now called first so that the invariant that there's never duplicate elements in the component list always holds, this fixes #7668. addComponent(c, index) now has a piece of additional logic handling the case when a component is moved to an index *after* its current one.



svn changeset:22117/svn branch:6.7