Class AbstractTransactionalQuery should be public (#12285)
AbstractTransactionalQuery was not visible outside its package
(default visibility). Added `public`.

Making this class public avoids lots of boilerplate code when making
alternate implementations of a QueryDelegate.

A standard TableQuery does a lot of introspection on the DB schema. 
If the schema is known beforehand, alternate implementations may specify
this information explicitly, and inherit from AbstractTransactionalQuery
to reuse part of the implementation.

Change-Id: I7d6b399ed7c755941ee9be7931a2ba80e812b5ab