Remove out of order flush() (#12283)
Writer.flush() is specified as:
"Flushes the stream. If the stream has saved any characters from the various
write() methods in a buffer, write them immediately to their intended
destination. Then, if that destination is another character or byte stream,
flush it. Thus one flush() invocation will flush all the buffers in a chain
of Writers and OutputStreams."

Change-Id: Iecafdbb13a7a26a48300d2967c25c705f5bf1e38