Force connection to closed state on reconnect (#12853)
IE10+ does not call onabort when the streaming connection is aborted.
As a result, onclose is never called and the client side does not
change the connection state away from "connected". This results in
user actions not being queued on the client side, instead the client
tries to send them immediately, leading to an eternal spinner.

If this can be fixed in Atmosphere somehow, it should. This is a
workaround until then.

Change-Id: Icd1a86ac9775d1a7d9c0bc1dee0fd399e2bf2266