Avoid eagerly layouting from VScrollTable.updateFromUIDL (#13188)
Closing a modal sub-window at the same time when TreeTable item is
removed, caused the detached Window being re-opened by
WindowConnector.postLayout() call.

This change adds a check in postLayout: continue operation only if the
window is attached to DOM. Or else, log a warning message about the
invalid postLayout call. Another change is in TreeTableConnector and
VScrollTable to disallow Util.notifyParentOfSizeChange(Widget, boolean)
with a boolean 'false' argument, when rendering is in progress. 'false'
causes an immediate LayoutManager.layoutNow() call, which is the main
reason for this issue.

Change-Id: I6f3e331b0feff9e7814ae1d749f6f7812dcd49ac