Re-apply old change which was excluded from master
The merge script was faulty between May and June 2013 and merged changes
individually from both 7.0 and 7.1 to master. In the case that a commit from
7.0 was merged right before there was a "Merge: no" commit in 7.1, the changes
between the "Merge: no" commit and the previous merge point from 7.1 to master was
potentially never merged (the whole chain was merged as a no-op).

The only affected change seems to be

0a437a5 Revert "Replaced css inject hack in TestUtils with Page.Styles.add() #11798"
b008768 Implement parenthesis-handling fixes for Sass in Vaadin 7.1 (#12834)

The former one is re-applied in this commit, the latter one has been verified to be in
the master branch

Change-Id: If0cd75d75fdfc8893ac501aab2a0be1b55b09dcc