Move ActiveRowHandler to the data provider extension (#13334)
This makes GridConnector include information about what rows are cached
when more data is requested instead of the previous way of synchronizing
this separately every time a scroll event occurs. This new approach
makes it possible to have rows cached even if they are not in view. It
also improves performance since there's no need to do an RPC every time
scrolling changes what is visible.

Change-Id: Ibfe8a69586dfc397591f56efa8ef351e274f0116