Stop running TB2 tests on Safari 5 and Opera 12
While still supported, these browsers are completely outdated and the
tests no longer give us meaningful information. When TestBench 2 tests
are converted to TB4, they will no longer be run on either of these browser.

To improve the situation in the future, TB4 tests should be run on
Safari 7 and latest Opera

Change-Id: Ib49376e30e105aea89218aa352c44a4c3ec8778f