Allow for multiple license keys in a single license file #14408
License files can now contain multiple keys to support different
versions of a product. The keys are identified by the major version
number followed by an equals sign and the key for that version, e.g.

  1 = foo-bar-baz
  2 = baz-foo-bar

The license file can also contain a "default" license key, which
is used for all versions except for the explicitly defined ones, e.g.

  foo-bar-baz
  3 = baz-bar-foo
  4 = bar-baz-baz

Change-Id: Id07d22e9fdc44189c4298b634006cf0df128bfd9