Re-adding all rows in Table causes table to loose scroll position (#14581)
Fix: if to remove (container.removeAllItems()) and then to re-add(container.addAll(..)) the same collection in table container - > scroll position is not loosed now.

The hash code and scroll position of the old container is stored when a new container is added.
If the new container has the same hash code, we restore the scroll position.
The scroll position is not restored if another items added to the container.

Change-Id: I52a22c3c1c7b71f1b3447b9d592ab8fececd67b8