always retry a status code 0 uidl request (#15254)
A status code of 0 indicates that the request has not been handled by
the server because the connection was closed without sending response
headers.
This can happen on the browser side when the browser window is closed,
or it can happen on network equipment side under adverse quality of link
conditions.
Therefore vaadin should retry such a request once before popping up the
disconnect message.

Change-Id: I9524c97a3de8e4ac9499394a2a795b0441a36ead