Fix writing issues (#7749)
* A root component with @DesignRoot must always use its superclass for
default values. Otherwise the written design will be empty as the
design is read in the constructor.

* A component which is not the root must not write its component tree
if the component tree is generated in the constructor. This is a
simplification which should be good enough for most cases (can't add
children in constructor and also using addComponent - in this case
the component added with addComponent will not be written).

* Test cases for nested templates

Change-Id: I3a384d1d8654b9865a3a790ebeb055a300a62135