Fix grid row borders and cell widget vertical alignment (#16200)
Previously the top-border for the first body row in the grid was
hidden, which caused it to blink during scrolling, when the first-child
element is moved using transforms. Moving the border to the bottom of
the cell seems to have to negative side-effects.

Fixes vertical alignment for contained widget renderers. All DIV
elements are rendered as inline-blocks by default (same as all
.v-widget elements).

Change-Id: I277f88953069846adbc475366161aabc0df96e15