Grid columns can now be marked as non-editable (#16538)
Non-editable columns are not assigned editor fields. When the editor is active,
any non-editable content is not displayed (this should changein the future).
This is separate from setting the property or editor field read-only - in those
cases the field is still used to display the data which may not be desired and
will fail if there is no converter.

Also add Column.setEditorField(Field<?>) and the corresponding getter.

Change-Id: Ice17c357895cb63a8e1bfd6abaffc1d803399e98