Required indicator for Table now works as expected (#17294)
Table is a field (and one of the most powerful ones), but it cannot be
used as required, especially if one wants to use the built in "required
indicators" in Vaadin. Those only work if client side connector extends
AbstractFieldConnector so Table is now extending it as well and
(I expect that it ) needs to do the AbstractHasComponentsConnector
stuff internally.

Change-Id: I4ceb719366621053c7fe4c9691d624d61ed57567