Detect broken Profiler.isEnabled() inlining
The mechanism that makes calls to e.g. Profiler.start() completely
disappear if vaadin.profiler is set to false is quite fragile.
Furhtermore, the implementation pattern has some parts that look like
potential optimizations even though the end result would be the
opposite. This patch adds a test that will fail if that kind of change
has accidentally been introduced.

Change-Id: Ic948feb4d3b73120dc3f245bcc130af16bfd32a9