TwinColSelect with new databinding API
Removes feature for adding new items.
Introduces a AbstractMultiSelect-abstraction layer,
which is used in server side by TwinColSelect & CheckBoxGroup and
on client side only TwinColSelect for now. Plan is to use it for
ListSelect too.

Further improvement would be to make AbstractMultiSelect use
SelectionModel that extends AbstractSelectionModel and is thus used
as an extension both as client & server side.

Updates to JUnit 4.12 for easier use of @Parameterized test..

Change-Id: I64258c2229b9514d382693748e2ca562a1e448d4